Converted Draggable component
